+++ packages/thumbprint-react/CHANGELOG.md
@@ -6,6 +6,10 @@
 
 -   [Patch] Add note in `dataTest` deprecated documentation telling developers to use `dataTestId` instead.
 
+### Added
+
+-   [Minor] Add missing `dataTestId` prop to `Link` and `ThemedLink` components. (#724)
+
 ## 14.2.0 - 2020-10-21
 
 ### Added